+++ src/table/simple-table.tsx
@@ -64,8 +64,6 @@ type Props = {
   hasColumnVisibilityChooser?: boolean;
   /** When false, turns off text wrapping for table headers */
   wrapHeader?: boolean;
-  /** Add padding to the left of the table in pixels. This is most useful when the table is not selectable */
-  paddingLeft?: number;
 };
 
 /** Provides a simple table for displaying data, with the ability to opt into additional features. */