+++ src/table/utilities/checkbox-column.tsx
@@ -1,6 +1,7 @@
 import Checkbox from "@material-ui/core/Checkbox/Checkbox";
 import React from "react";
 import classNames from "classnames";
+import styles from "../table.module.scss";
 import {
   InternalPagedTableColumn,
   PagedTableCell,
@@ -9,7 +10,51 @@ import {
 
 export const getCheckboxColumn = (
   selectionProps: SelectionProps,
+  shouldAdjustTable: boolean,
 ): InternalPagedTableColumn => {
+  if (shouldAdjustTable) {
+    return {
+      accessor: "COLUMN_SELECT",
+      headerClassName: styles.checkBoxColumn,
+      className: styles.checkBoxColumn,
+      Header: (
+        <Checkbox
+          classes={{
+            root: classNames({
+              [styles.checkBox]: shouldAdjustTable,
+            }),
+          }}
+          inputProps={{
+            // @ts-ignore: data-testid is nnt assignable.
+            "data-testid": "hidden-table-checkbox-header",
+          }}
+        />
+      ),
+      Cell: ({ rowId }: PagedTableCell<any>) => {
+        const testId = String(rowId)
+          .toLowerCase()
+          .replace(" ", "-");
+        return (
+          <Checkbox
+            id={String(rowId)}
+            classes={{
+              root: classNames({
+                [styles.checkBox]: shouldAdjustTable,
+              }),
+            }}
+            inputProps={{
+              // @ts-ignore: data-testid is not assignable.
+              "data-testid": `hidden-table-checkbox-cell-${testId}`,
+            }}
+          />
+        );
+      },
+      isVisible: true,
+      index: -1,
+      exportHeaderName: "",
+      exportAccessor: "",
+    };
+  }
   const { data, selectedRows, idKey, onSelect } = selectionProps;
   if (selectedRows === undefined || selectedRows === null) {
     throw new Error("selectedRows is missing from props");
@@ -17,18 +62,16 @@ export const getCheckboxColumn = (
   const ids = data.map((instance, index) =>
     idKey ? instance[idKey] : `${index}`,
   );
-  // FIXME (nsawas/jsingh): allAreChecked will only work as expected if API is retrieving data using count parameter.
-  const allAreChecked =
-    selectedRows.length === data.length && data.length !== 0;
-  const someAreChecked = selectedRows.length > 0;
+  // FIXME (nsawas): allAreSelected will only work as expected if API is retrieving data using count parameter.
+  const allAreSelected = ids.every(id => selectedRows.includes(id));
+  const someAreSelected = selectedRows.length > 0 && !allAreSelected;
 
   const selectAll = (): void => {
-    const areAllSelected = ids.every(id => selectedRows.includes(id));
-    if (areAllSelected) {
+    if (someAreSelected || allAreSelected) {
       onSelect && onSelect([]);
-      return;
+    } else {
+      onSelect && onSelect(ids);
     }
-    onSelect && onSelect(ids);
   };
 
   const selectItem = (rowId: number | string): void => {
@@ -40,22 +83,24 @@ export const getCheckboxColumn = (
     }
     onSelect && onSelect(newSelection);
   };
-
   return {
     accessor: "COLUMN_SELECT",
+    className: styles.checkBoxColumn,
     Header: (
       <Checkbox
-        className={classNames({
-          active: allAreChecked,
-          indeterminate: someAreChecked,
-        })}
+        classes={{
+          root: classNames({
+            active: allAreSelected,
+            indeterminate: someAreSelected,
+          }),
+        }}
         inputProps={{
           // @ts-ignore: data-testid is nnt assignable.
           "data-testid": "table-checkbox-header",
         }}
         color="primary"
-        checked={someAreChecked || allAreChecked}
-        indeterminate={someAreChecked && !allAreChecked}
+        checked={allAreSelected || !!someAreSelected}
+        indeterminate={someAreSelected}
         onChange={selectAll}
       />
     ),
@@ -66,7 +111,11 @@ export const getCheckboxColumn = (
       return (
         <Checkbox
           id={String(rowId)}
-          className={classNames({ active: selectedRows.includes(rowId) })}
+          classes={{
+            root: classNames({
+              active: selectedRows.includes(rowId),
+            }),
+          }}
           inputProps={{
             // @ts-ignore: data-testid is not assignable.
             "data-testid": `table-checkbox-cell-${testId}`,